requirements.txt
google-cloud-vision==0.34.0
google-cloud-speech==0.36.0
google-cloud-texttospeech==0.2.0
google-cloud-datastore==1.7.1
Flask==1.0.2
Flask-HTTPAuth==3.2.4